fix(import): clamp Marinara lorebook entry role to the system/user/assistant union
The Marinara native importer wrote the lorebook entry role verbatim
(role: (e.role as any) ?? "system") and the bulk-import path (bulkCreateEntries
-> createEntry) does no zod parse, so an arbitrary string from a hand-crafted
.marinara.json persisted as the entry role and later propagated into a provider
PromptMessage role. The ST lorebook importer already normalizes the identical
field via resolveRole, and the manual entry-create routes run
createLorebookEntrySchema.parse() — the Marinara bulk path was the one site
that skipped the clamp.

Extract the ST importer's resolveRole into a shared resolveLorebookEntryRole
helper (param widened to unknown) and call it from both importers, so the
Marinara path clamps out-of-enum roles to "system" while legitimate roles
("system"/"user"/"assistant" and ST numeric 0/1/2) round-trip unchanged.

Refs #2530
